refactor(web/engine): starts DOM pre-processor
fix(web): Ensures dev resources continue working
